rename: CreateBand 폴더 이동
